feat: svg caching
The `preloadSvg` function loads and caches the given <svg> source in memory until `unloadSvg` is called for the same <svg> source.

Use the returned "cache key" by passing it as the `cacheKey` prop of an <SVGKView> component. When the `cacheKey` prop is defined, the `data` and `source` props can both be undefined.

Caching an <svg> source by URL allows any <SVGKView> component with the same URL to share the <svg> data, even with an undefined `cacheKey` prop.

Raw <svg> strings can also be cached, which stores the parsed <svg> data in memory.

The `useSvg` hook handles the `preloadSvg` and `unloadSvg` calls for you. A "cache key" is returned to use as the `cacheKey` prop of an <SVGKView> component.