fix: avoid resizing an old image view
Right before the "_imageView" is replaced, it was being resized by a call to "setFrame:". When the old image has a different size than the new image, this results in the old image resizing to the new image's size for a single frame before the new image was rendered instead, which caused an undesirable visual jump.